Various fixes.
- Restored code saving Store to localStorage that I accidentally removed
in previous commit.
- Global storeAuthUser now returns a promise that resolves when the user data
is fetched from the backend.
- Fixed various async issues with getting an authed user into global Store.
1. On App startup, App render should wait for getAuthUser() to complete before
rendering if a token is present in localStorage and the global Store has no
user in state.
2. Header can't assume a user is always available because it needs to work
even when the user isn't logged in which also means it can't just wait
for user auth to complete, so I wrapped the user data fetching in conditionals.
3. Register now properly chains on authUser so that it waits for the user
data to populate into Store.state.
4. User profile only renders when the user data is ready (from Register's setState)
so the user data just needs to be fetched during render. Doing it at file
scope is too early because the user data hasn't populated to Store when files
are loaded.